requirements.txt
django
Pillow
requests
djangorestframework